redo self.special_qualities as list, not str
The monster_object was edited so that the attribute
special_qualities is now a list.  It was a string.
A list allows for checking to see if items are already on the list
as well as other list operations.

This is to pave the way for applying some special qualities
as part of the celestial template (or infernal template)
to existing special qualities.