modify sq with celestial or infernal template
The +good (or +g or +celestial) modifier adds the celstial template
to *some* summoned creatures, which allows them to smite evil for
damage based on Hit Dice.  Not all summoned creatures take this
template: demons are one example, elementals another.

The +evil (or +e or +infernal) modifier adds the infernal template
to *some* summoned creatures, which allows them to smite good for
damage based on Hit Dice.  Not all summoned creatures take this
template: demons are one example, elementals another. archons a third.

Also adds a unit test or two to check for how this is built.

Requires the element
  <takes_c_or_i_template/>
in each <monster> element in the tree.  There are probably other
ways of doing this, but I wanted to try searching for
the presence of a boolean element (either it's there and True,
or it is not so False).